Refactor static logger initialization
